[WIP] remove models, create new controllers. ajax validation
