+++ Classes/Domain/Repository/SubscriptionRepository.php
@@ -1,0 +1,15 @@
+<?php
+declare(strict_types=1);
+
+namespace Pixelant\PxaNewsletterSubscription\Domain\Repository;
+
+use TYPO3\CMS\Extbase\Persistence\Repository;
+
+/**
+ * Class SubscriptionRepository
+ * @package Pixelant\PxaNewsletterSubscription\Domain\Repository
+ */
+class SubscriptionRepository extends Repository
+{
+
+}
