+++ Classes/Utility/FrontendUserStorageUtility.php
@@ -121,7 +121,7 @@ class FrontendUserStorageUtility extends \Pixelant\PxaNewsletterSubscription\Uti
 
         // Signal slot for after fe_user creation
         $this->signalSlotDispatcher->dispatch(
-            \Pixelant\PxaNewsletterSubscription\Controller\NewsletterSubscriptionController::class,
+            \Pixelant\PxaNewsletterSubscription\Controller\NewsletterSubscriptionControllerOld::class,
             'afterFeUserCreation',
             [$frontendUser, $this]
         );
